+++ packages/orchestration/src/orchestration-types.ts
@@ -1,6 +1,8 @@
 import type { ChatModel } from './model-types.js';
 import type {
   ChatMessages,
+  DataRepositoryType,
+  DocumentGroundingFilter,
   FilteringModuleConfig,
   GroundingModuleConfig,
   MaskingModuleConfig,
@@ -55,4 +57,43 @@ export interface OrchestrationModuleConfig {
    * Grounding module configuraton.
    */
   grounding?: GroundingModuleConfig;
+}
+
+/**
+ * Represents a filter configuration for the Document Grounding Service.
+ *
+ */
+export type DocumentGroundingServiceFilter = Pick<
+  DocumentGroundingFilter,
+  | 'id'
+  | 'search_config'
+  | 'data_repositories'
+  | 'data_repository_metadata'
+  | 'document_metadata'
+  | 'chunk_metadata'
+> & {
+  /**
+   * Defines the type of data repository.
+   * If not set, the default value is 'vector'.
+   */
+  data_repository_type?: DataRepositoryType;
+};
+
+/**
+ * Represents the configuration for the Document Grounding Service.
+ */
+export interface DocumentGroundingServiceConfig {
+  /**
+   * Define the filters to apply during the grounding process.
+   */
+  filters?: DocumentGroundingServiceFilter[];
+  /**
+   * Contains the input parameters used for grounding input questions.
+   */
+  input_params: string[];
+  /**
+   * Parameter name used for grounding output.
+   * @example "groundingOutput"
+   */
+  output_param: string;
 }
